+++ autoendpoint/src/db/client.rs
@@ -3,6 +3,9 @@ use async_trait::async_trait;
 use autopush_common::db::UserRecord;
 use autopush_common::notification::Notification;
 use std::collections::HashSet;
+use std::env;
+use std::sync::Arc;
+use uuid::Uuid;
 
 use uuid::Uuid;