+++ autoendpoint/src/extractors/routers.rs
@@ -14,7 +14,7 @@ use std::str::FromStr;
 use std::sync::Arc;
 
 /// Valid `DynamoDbUser::router_type` values
-#[derive(Copy, Clone, Debug, PartialEq)]
+#[derive(Copy, Clone, Debug, Eq, PartialEq)]
 #[allow(clippy::upper_case_acronyms)]
 pub enum RouterType {
     WebPush,