+++ tests/test_integration_all_rust.py
@@ -2,12 +2,14 @@
 Rust Connection and Endpoint Node Integration Tests
 """
 
+from base64 import urlsafe_b64encode
 import copy
 import json
 import logging
 import os
 import random
 import signal
+import string
 import socket
 import subprocess
 
@@ -70,6 +72,7 @@ MOCK_SERVER_THREAD = None
 CN_QUEUES = []
 EP_QUEUES = []
 STRICT_LOG_COUNTS = True
+RUST_LOG = "autoendpoint=debug,autopush_rs=debug,autopush_common=debug,error"
 
 
 def get_free_port():
@@ -96,8 +99,13 @@ CONNECTION_CONFIG = dict(
     statsd_host="",
     router_tablename=ROUTER_TABLE,
     message_tablename=MESSAGE_TABLE,
+<<<<<<< HEAD
     crypto_keys="[{}]".format(CRYPTO_KEY),
     auto_ping_interval=60.0,
+=======
+    crypto_key="[{}]".format(CRYPTO_KEY),
+    auto_ping_interval=30.0,
+>>>>>>> e0c370ae7d00405adf4f96ada9e94afe3208b76b
     auto_ping_timeout=10.0,
     close_handshake_timeout=5,
     max_connections=5000,
@@ -247,7 +255,7 @@ keyid="http://example.org/bob/keys/123";salt="XZwpw6o37R-6qoZjw6KwAw=="\
     def send_notification(self, channel=None, version=None, data=None,
                           use_header=True, status=None, ttl=200,
                           timeout=0.2, vapid=None, endpoint=None,
-                          topic=None):
+                          topic=None, headers=None):
         if not channel:
             channel = random.choice(self.channels.keys())
 
@@ -285,6 +293,7 @@ keyid="http://example.org/bob/keys/123";salt="XZwpw6o37R-6qoZjw6KwAw=="\
         status = status or 201
 
         log.debug("%s body: %s", method, body)
+        log.debug("  headers: %s", headers)
         http.request(method, url.path.encode("utf-8"), body, headers)
         resp = http.getresponse()
         log.debug("%s Response (%s): %s", method, resp.status, resp.read())
@@ -333,7 +342,8 @@ keyid="http://example.org/bob/keys/123";salt="XZwpw6o37R-6qoZjw6KwAw=="\
             result = json.loads(d)
             assert result.get("messageType") == "broadcast"
             return result
-        except Exception:  # pragma: nocover
+        except Exception as ex:  # pragma: nocover
+            log.error("Error: {}".format(ex))
             return None
         finally:
             self.ws.settimeout(orig_timeout)
@@ -509,6 +519,7 @@ def get_rust_binary_path(binary):
 def write_config_to_env(config, prefix):
     for key, val in config.items():
         new_key = prefix + key
+        log.debug("#### {} => {}".format(new_key, val))
         os.environ[new_key.upper()] = str(val)
 
 
@@ -559,7 +570,11 @@ def setup_mock_server():
 def setup_connection_server(connection_binary):
     global CN_SERVER
 
-    write_config_to_env(CONNECTION_CONFIG, "autopush_")
+    # NOTE:
+    # due to a change in Config, autopush uses a double
+    # underscore as a separator (e.g. "AUTOEND__FCM__MIN_TTL" ==
+    # `settings.fcm.min_ttl`)
+    write_config_to_env(CONNECTION_CONFIG, "autopush__")
     cmd = [connection_binary]
     try:
         CN_SERVER = subprocess.Popen(
@@ -578,7 +593,7 @@ def setup_connection_server(connection_binary):
 def setup_megaphone_server(connection_binary):
     global CN_MP_SERVER
 
-    write_config_to_env(MEGAPHONE_CONFIG, "autopush_")
+    write_config_to_env(MEGAPHONE_CONFIG, "autopush__")
     cmd = [connection_binary]
     try:
         CN_MP_SERVER = subprocess.Popen(cmd, shell=True, env=os.environ)
@@ -590,8 +605,11 @@ def setup_endpoint_server():
     global EP_SERVER
 
     # Set up environment
-    os.environ["RUST_LOG"] = "trace"
-    write_config_to_env(ENDPOINT_CONFIG, "autoend_")
+    # NOTE:
+    # due to a change in Config, autoendpoint uses a double
+    # underscore as a separator (e.g. "AUTOEND__FCM__MIN_TTL" ==
+    # `settings.fcm.min_ttl`)
+    write_config_to_env(ENDPOINT_CONFIG, "autoend__")
 
     # Run autoendpoint
     cmd = [get_rust_binary_path("autoendpoint")]
@@ -612,7 +630,7 @@ def setup_endpoint_server():
 
 def setup_module():
     global CN_SERVER, CN_QUEUES, CN_MP_SERVER, MOCK_SERVER_THREAD, \
-        STRICT_LOG_COUNTS
+        STRICT_LOG_COUNTS, RUST_LOG
 
     if "SKIP_INTEGRATION" in os.environ:  # pragma: nocover
         raise SkipTest("Skipping integration tests")
@@ -627,6 +645,7 @@ def setup_module():
 
     setup_mock_server()
 
+    os.environ["RUST_LOG"] = RUST_LOG
     connection_binary = get_rust_binary_path("autopush_rs")
     setup_connection_server(connection_binary)
     setup_megaphone_server(connection_binary)
@@ -661,6 +680,7 @@ class TestRustWebPush(unittest.TestCase):
 
     @inlineCallbacks
     def quick_register(self, sslcontext=None):
+        print("#### Connecting to ws://localhost:{}/".format(CONNECTION_PORT))
         client = Client("ws://localhost:{}/".format(CONNECTION_PORT),
                         sslcontext=sslcontext)
         yield client.connect()
@@ -1212,6 +1232,26 @@ class TestRustWebPush(unittest.TestCase):
         yield client.ack(result3["channelID"], result3["version"])
 
         yield self.shut_down(client)
+
+    @inlineCallbacks
+    def test_big_message(self):
+        """Test that we accept a large message.
+
+        Using pywebpush I encoded a 4096 block
+        of random data into a 4216b block. B64 encoding that produced a block that was
+        5624 bytes long. We'll skip the binary bit for a 4216 block of "text" we then
+        b64 encode to send.
+        """
+        import base64;
+        client = yield self.quick_register()
+        data = base64.urlsafe_b64encode(
+            ''.join(random.choice(string.ascii_letters+string.digits+string.punctuation)
+            for _ in xrange(0, 4216))
+            )
+        result = yield client.send_notification(data=data)
+        dd = result.get("data")
+        dh = base64.b64decode(dd + "==="[:len(dd) % 4])
+        assert dh == data
 
     # Need to dig into this test a bit more. I'm not sure it's structured correctly
     # since we resolved a bug about returning 202 v. 201, and it's using a dependent